feat(refactor): replace settings tab with More screen
Closes #113

- Add MoreScreen with Analytics, Goals, Backup Data options and a gear FAB
- Extract GoalsScreen and BackupScreen from SettingsScreen
- Slim SettingsScreen to language, appearance and units
- Move screens into app/(tabs) and register hidden routes so bottom nav stays visible
- Update i18n keys (en + de) and tabs layout icon
