+++ app/(tabs)/analytics.tsx
@@ -1,0 +1,1 @@
+export { default } from "../../src/features/analytics/AnalyticsScreen";
