+++ app/(tabs)/backup.tsx
@@ -1,0 +1,1 @@
+export { default } from "@/src/features/settings/BackupScreen";
