+++ app/(tabs)/goals.tsx
@@ -1,0 +1,1 @@
+export { default } from "@/src/features/settings/GoalsScreen";
