--- app/analytics.tsx
@@ -1,1 +1,0 @@
-export { default } from "../src/features/analytics/AnalyticsScreen";
